requirements_gpu.txt
h5py==2.10
matplotlib==3.1
numpy==1.18
opencv-python==4.1.2
Pillow==7.0
scikit-image==0.16
scikit-learn==0.22
scipy==1.4
tensorflow-gpu==1.13